cosmetics: STREAMS.cc, reformat some examples, MPFR_interface.h, exp_log.cc, sin_cos.cc, pi_ln2.cc, sqrt.cc

+++ examples/test_commandline.cc
@@ -3,20 +3,25 @@
 
 using namespace iRRAM;
 
-template int iRRAM::iRRAM_exec <int,int,char **> (int (*) (const int &, char **const &),const int &, char **const &);
+template int iRRAM::iRRAM_exec<int, int, char **>(int (*)(const int &,
+                                                          char ** const &),
+                                                  const int &, char ** const &);
 
-int iRRAM_compute(const int &argc, char **const &argv) {
+int iRRAM_compute(const int & argc, char ** const & argv)
+{
 
-REAL x1,x2;
-if (argc >=2 ) x1=argv[1];
-if (argc >=3 ) x2=argv[2];
-cout <<x1<<"\n"<<x2<<"\n";
+	REAL x1, x2;
+	if (argc >= 2)
+		x1 = REAL(argv[1]);
+	if (argc >= 3)
+		x2 = REAL(argv[2]);
+	cout << x1 << "\n" << x2 << "\n";
 
-return 0;
+	return 0;
 }
 
-int main (int argc,char **argv)
+int main(int argc, char ** argv)
 {
-iRRAM_initialize(argc,argv);
-return iRRAM_exec(iRRAM_compute,argc,argv);
+	iRRAM_initialize(argc, argv);
+	return iRRAM_exec(iRRAM_compute, argc, argv);
 }
